client/python/requirements.txt
aiohttp>=3.8.0
sseclient-py>=1.7.2
requests>=2.28.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-timeout>=2.1.0 